+++ tfjs-core/src/tests.ts
@@ -105,6 +105,7 @@ import './ops/max_pool_3d_test';
 import './ops/max_pool_test';
 import './ops/max_pool_with_argmax_test';
 import './ops/mean_squared_error_test';
+import './ops/moments_test';
 import './ops/moving_average_test';
 import './ops/multi_rnn_cell_test';
 import './ops/multinomial_test';
@@ -115,6 +116,7 @@ import './ops/one_hot_test';
 import './ops/operation_test';
 import './ops/pad_test';
 import './ops/pool_test';
+import './ops/prod_test';
 import './ops/qr_test';
 import './ops/rand_test';
 import './ops/random_gamma_test';
